# Fix `function-calc-no-unspaced-operator` performance