--- a/.changeset/wet-countries-think.md
+++ b/.changeset/wet-countries-think.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `function-calc-no-unspaced-operator` performance