--- a/lib/rules/function-calc-no-unspaced-operator/index.js
+++ b/lib/rules/function-calc-no-unspaced-operator/index.js
@@ -48,6 +48,8 @@ const rule = (primary, _secondaryOptions, context) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!OPERATOR_REGEX.test(getDeclarationValue(decl))) return;
+
 			let needsFix = false;
 			const valueIndex = declarationValueIndex(decl);
 			const parsedValue = valueParser(getDeclarationValue(decl));