--- a/lib/rules/function-calc-no-unspaced-operator/index.js
+++ b/lib/rules/function-calc-no-unspaced-operator/index.js
@@ -48,11 +48,13 @@ const rule = (primary, _secondaryOptions, context) => {
 		}
 
 		root.walkDecls((decl) => {
-			if (!OPERATOR_REGEX.test(getDeclarationValue(decl))) return;
+			const value = getDeclarationValue(decl);
+
+			if (!OPERATOR_REGEX.test(value)) return;
 
 			let needsFix = false;
 			const valueIndex = declarationValueIndex(decl);
-			const parsedValue = valueParser(getDeclarationValue(decl));
+			const parsedValue = valueParser(value);
 
 			/**
 			 * @param {import('postcss-value-parser').Node} operatorNode