<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See : https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

No, it's self-explanatory.

----

before :

```
Warnings: 0
Mean: 90.56110877777778 ms
Deviation: 11.54085198893875 ms
```

after : 

```
Warnings: 0
Mean: 83.60294444444445 ms
Deviation: 10.305877029720664 ms
```