@romainmenke 2023-06-17
GitHub seems to be having some issues.
It didn't apply the review suggestions correctly and pushed to this branch don't show up.

I am giving it some time before make further changes.